ADD: Refactor input encoding structure
