+++ src/encoding/input_encoders/audio_input_encoders.py
@@ -1,0 +1,18 @@
+from .input_encoder import AbstractInputEncoder
+
+
+import torch
+
+
+class AudioInputEncoder(AbstractInputEncoder):
+    def __init__(self):
+        super().__init__()
+
+    def fit(self, inputs):
+        pass
+
+    def transform(self, data):
+        pass
+
+    def collate_fn(self, batch):
+        pass
